moved redirect for not-logged-in-users to header.php
As theme I use a child of the Origami theme (by SiteOrigin: https://siteorigin.com/theme/origami/ ). To adapt the theme, I made a child theme (that is what this git is all about).

For the redirect, I changed the header.php, but I still require the original theme's header. In most documentations, you will find "make a copy of the header.php to your child theme and edit from there", but I prefer to make a copy, rename it and "require" it through the new header.php . That way, it can still work with updated versions of the theme, without too much hassle: just make a copy of the parent's header.php and rename it.