dirty hack:
when the preferred_language is changed in the page-profiel-ophaalpunt.php ,
that change is not yet put in the DB before user_language() is called.

So when user_language() is called, it is not aware of the recent language change.

This means that the first reload of the profile-page will still be in the previous preferred_language, probably confusing the user.

By doing this dirty hack, the code knows that the preferred_language has been changed and can translate the page accordingly.