Dirty Hack: make userIDs of Administrators into GLOBAL variables
so that we can exclude the locations (that are linked to them) from the list presented in "register new user"

We want to show only those locations that have not been linked to a user yet. But what if a location has been linked to one of the administrators of MyRecy for testpurposes? We should not exclude that one (unless it has been linked to a regular user as well)

so:

use 'SELECT ophaalpunten.* FROM ophaalpunten ORDER BY naam'
if you want to show ALL locations

use 'SELECT ophaalpunten.* FROM ophaalpunten WHERE ophaalpunten.id not in (select `ophaalpunt_id` from wordpress_link)  ORDER BY naam'
if you only want to show the locations that have not been linked to a user yet

(WATCH OUT, what if a location ("ophaalpunt") has been linked to an ADMINISTRATOR, then it will also not show!
--> for this we can use:
SELECT ophaalpunten.* FROM ophaalpunten WHERE ophaalpunten.id not in (select `ophaalpunt_id` from wordpress_link where wordpress_userid <> 1 AND wordpress_userid <> 4)  ORDER BY naam

ugggh....